Add support for AdaFruit CC3000
