+++ examples/Web_HelloWorld/Files/HelloWorld.h
@@ -1,0 +1,19 @@
+//
+//  ___FILENAME___
+//
+//  Project ___PROJECTNAME___
+//  Created by ___FULLUSERNAME___ on ___DATE___
+//  Copyright (c) ___YEAR___ ___ORGANIZATIONNAME___
+//
+
+#ifndef ___PACKAGENAME_____h
+#define ___PACKAGENAME_____h
+
+#include  "Arduino.h"
+
+
+
+
+
+
+#endif
