+++ examples/Web_HelloWorld/HelloWorld.xcodeproj/project.xcworkspace/xcshareddata/HelloWorld.xccheckout
@@ -1,0 +1,41 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
+<plist version="1.0">
+<dict>
+	<key>IDESourceControlProjectFavoriteDictionaryKey</key>
+	<false/>
+	<key>IDESourceControlProjectIdentifier</key>
+	<string>E67FDCC6-7EC4-4C4B-9109-5BD111040CA6</string>
+	<key>IDESourceControlProjectName</key>
+	<string>HelloWorld</string>
+	<key>IDESourceControlProjectOriginsDictionary</key>
+	<dict>
+		<key>F79E376B-0C4C-4FB3-A56C-DFFCD607AA2A</key>
+		<string>ssh://github.com/corbinstreehouse/Webduino.git</string>
+	</dict>
+	<key>IDESourceControlProjectPath</key>
+	<string>examples/Web_HelloWorld/HelloWorld.xcodeproj/project.xcworkspace</string>
+	<key>IDESourceControlProjectRelativeInstallPathDictionary</key>
+	<dict>
+		<key>F79E376B-0C4C-4FB3-A56C-DFFCD607AA2A</key>
+		<string>../../../..</string>
+	</dict>
+	<key>IDESourceControlProjectURL</key>
+	<string>ssh://github.com/corbinstreehouse/Webduino.git</string>
+	<key>IDESourceControlProjectVersion</key>
+	<integer>110</integer>
+	<key>IDESourceControlProjectWCCIdentifier</key>
+	<string>F79E376B-0C4C-4FB3-A56C-DFFCD607AA2A</string>
+	<key>IDESourceControlProjectWCConfigurations</key>
+	<array>
+		<dict>
+			<key>IDESourceControlRepositoryExtensionIdentifierKey</key>
+			<string>public.vcs.git</string>
+			<key>IDESourceControlWCCIdentifierKey</key>
+			<string>F79E376B-0C4C-4FB3-A56C-DFFCD607AA2A</string>
+			<key>IDESourceControlWCCName</key>
+			<string>Webduino</string>
+		</dict>
+	</array>
+</dict>
+</plist>
