+++ examples/Web_HelloWorld/HelloWorld.xcodeproj/project.xcworkspace/xcshareddata/HelloWorld2.xccheckout
@@ -1,0 +1,41 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
+<plist version="1.0">
+<dict>
+	<key>IDESourceControlProjectFavoriteDictionaryKey</key>
+	<false/>
+	<key>IDESourceControlProjectIdentifier</key>
+	<string>593C39BD-5E96-4741-94D1-9EEF465CC240</string>
+	<key>IDESourceControlProjectName</key>
+	<string>HelloWorld2</string>
+	<key>IDESourceControlProjectOriginsDictionary</key>
+	<dict>
+		<key>F79E376B-0C4C-4FB3-A56C-DFFCD607AA2A</key>
+		<string>ssh://github.com/corbinstreehouse/Webduino.git</string>
+	</dict>
+	<key>IDESourceControlProjectPath</key>
+	<string>examples/Web_HelloWorld/HelloWorld2.xcodeproj/project.xcworkspace</string>
+	<key>IDESourceControlProjectRelativeInstallPathDictionary</key>
+	<dict>
+		<key>F79E376B-0C4C-4FB3-A56C-DFFCD607AA2A</key>
+		<string>../../../..</string>
+	</dict>
+	<key>IDESourceControlProjectURL</key>
+	<string>ssh://github.com/corbinstreehouse/Webduino.git</string>
+	<key>IDESourceControlProjectVersion</key>
+	<integer>110</integer>
+	<key>IDESourceControlProjectWCCIdentifier</key>
+	<string>F79E376B-0C4C-4FB3-A56C-DFFCD607AA2A</string>
+	<key>IDESourceControlProjectWCConfigurations</key>
+	<array>
+		<dict>
+			<key>IDESourceControlRepositoryExtensionIdentifierKey</key>
+			<string>public.vcs.git</string>
+			<key>IDESourceControlWCCIdentifierKey</key>
+			<string>F79E376B-0C4C-4FB3-A56C-DFFCD607AA2A</string>
+			<key>IDESourceControlWCCName</key>
+			<string>Webduino</string>
+		</dict>
+	</array>
+</dict>
+</plist>
