+++ examples/Web_HelloWorld/HelloWorld.xcodeproj/project.xcworkspace/xcshareddata/LEDCyrWheel.xccheckout
@@ -1,0 +1,89 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
+<plist version="1.0">
+<dict>
+	<key>IDESourceControlProjectFavoriteDictionaryKey</key>
+	<false/>
+	<key>IDESourceControlProjectIdentifier</key>
+	<string>B0A14392-E54A-4AB1-AF39-705524C984E1</string>
+	<key>IDESourceControlProjectName</key>
+	<string>LEDCyrWheel</string>
+	<key>IDESourceControlProjectOriginsDictionary</key>
+	<dict>
+		<key>072E38A3-980A-451E-8E79-20BB7C84B2F5</key>
+		<string>https://github.com/adafruit/RTClib.git</string>
+		<key>2C1A2DF8-D38E-419B-8C3D-375B2F36F9DD</key>
+		<string>https://github.com/adafruit/SD.git</string>
+		<key>3FC67D25-67C3-4230-962E-0F9C6AA16DDB</key>
+		<string>https://github.com/corbinstreehouse/ElithionBMSCanBus.git</string>
+		<key>F6500F34-56E8-4B4C-A916-F8DE34822AAF</key>
+		<string>https://github.com/corbinstreehouse/ElithionBMSDisplay.git</string>
+		<key>F75E2DC4-41EC-40A9-A96A-322B59982638</key>
+		<string>https://github.com/corbinstreehouse/UTFT_Adafruit_22TFT</string>
+	</dict>
+	<key>IDESourceControlProjectPath</key>
+	<string>LEDCyrWheel.xcodeproj/project.xcworkspace</string>
+	<key>IDESourceControlProjectRelativeInstallPathDictionary</key>
+	<dict>
+		<key>072E38A3-980A-451E-8E79-20BB7C84B2F5</key>
+		<string>../../../../../Documents/Arduino/Libraries/RTClib</string>
+		<key>2C1A2DF8-D38E-419B-8C3D-375B2F36F9DD</key>
+		<string>../../../../../Documents/Arduino/Libraries/SD</string>
+		<key>3FC67D25-67C3-4230-962E-0F9C6AA16DDB</key>
+		<string>../../../../../Documents/Arduino/Libraries/Canbus</string>
+		<key>F6500F34-56E8-4B4C-A916-F8DE34822AAF</key>
+		<string>../..</string>
+		<key>F75E2DC4-41EC-40A9-A96A-322B59982638</key>
+		<string>../../../../../Documents/Arduino/Libraries/UTFT</string>
+	</dict>
+	<key>IDESourceControlProjectURL</key>
+	<string>https://github.com/corbinstreehouse/ElithionBMSDisplay.git</string>
+	<key>IDESourceControlProjectVersion</key>
+	<integer>110</integer>
+	<key>IDESourceControlProjectWCCIdentifier</key>
+	<string>F6500F34-56E8-4B4C-A916-F8DE34822AAF</string>
+	<key>IDESourceControlProjectWCConfigurations</key>
+	<array>
+		<dict>
+			<key>IDESourceControlRepositoryExtensionIdentifierKey</key>
+			<string>public.vcs.git</string>
+			<key>IDESourceControlWCCIdentifierKey</key>
+			<string>3FC67D25-67C3-4230-962E-0F9C6AA16DDB</string>
+			<key>IDESourceControlWCCName</key>
+			<string>Canbus</string>
+		</dict>
+		<dict>
+			<key>IDESourceControlRepositoryExtensionIdentifierKey</key>
+			<string>public.vcs.git</string>
+			<key>IDESourceControlWCCIdentifierKey</key>
+			<string>F6500F34-56E8-4B4C-A916-F8DE34822AAF</string>
+			<key>IDESourceControlWCCName</key>
+			<string>LEDCyrWheel</string>
+		</dict>
+		<dict>
+			<key>IDESourceControlRepositoryExtensionIdentifierKey</key>
+			<string>public.vcs.git</string>
+			<key>IDESourceControlWCCIdentifierKey</key>
+			<string>072E38A3-980A-451E-8E79-20BB7C84B2F5</string>
+			<key>IDESourceControlWCCName</key>
+			<string>RTClib</string>
+		</dict>
+		<dict>
+			<key>IDESourceControlRepositoryExtensionIdentifierKey</key>
+			<string>public.vcs.git</string>
+			<key>IDESourceControlWCCIdentifierKey</key>
+			<string>2C1A2DF8-D38E-419B-8C3D-375B2F36F9DD</string>
+			<key>IDESourceControlWCCName</key>
+			<string>SD</string>
+		</dict>
+		<dict>
+			<key>IDESourceControlRepositoryExtensionIdentifierKey</key>
+			<string>public.vcs.git</string>
+			<key>IDESourceControlWCCIdentifierKey</key>
+			<string>F75E2DC4-41EC-40A9-A96A-322B59982638</string>
+			<key>IDESourceControlWCCName</key>
+			<string>UTFT</string>
+		</dict>
+	</array>
+</dict>
+</plist>
