+++ examples/Web_HelloWorld/HelloWorld.xcodeproj/project.xcworkspace/xcshareddata/LEDDigitalCyrWheel.xccheckout
@@ -1,0 +1,113 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
+<plist version="1.0">
+<dict>
+	<key>IDESourceControlProjectFavoriteDictionaryKey</key>
+	<false/>
+	<key>IDESourceControlProjectIdentifier</key>
+	<string>23A547CD-4F70-441A-BCF7-0DC6FE479AA2</string>
+	<key>IDESourceControlProjectName</key>
+	<string>LEDDigitalCyrWheel</string>
+	<key>IDESourceControlProjectOriginsDictionary</key>
+	<dict>
+		<key>31D19277-93C1-4B36-8D47-39BAE008E6C8</key>
+		<string>https://github.com/adafruit/Adafruit_NeoPixel.git</string>
+		<key>581ADA72-D596-47B7-96A2-A3DFDF0C6864</key>
+		<string>https://github.com/corbinstreehouse/ElithionBMSCanBus.git</string>
+		<key>7B373F31-FC1C-4563-BCFF-DFA44AB4C0D0</key>
+		<string>https://github.com/adafruit/SD.git</string>
+		<key>7B5D90EB-E67D-4F4F-B891-A4D47362BD97</key>
+		<string>https://github.com/corbinstreehouse/LEDDigitalCyrWheel.git</string>
+		<key>B998CEF2-BA3A-4B9A-861D-23AB57E81EA4</key>
+		<string>https://github.com/FastLED/FastLED.git</string>
+		<key>BB79AA8B-9CED-45F6-9A86-44FF4DDD236C</key>
+		<string>https://github.com/adafruit/RTClib.git</string>
+		<key>D49A604E-E9FF-408A-B3BC-33F3DEF41E6B</key>
+		<string>https://github.com/corbinstreehouse/UTFT_Adafruit_22TFT</string>
+	</dict>
+	<key>IDESourceControlProjectPath</key>
+	<string>LEDDigitalCyrWheel.xcodeproj/project.xcworkspace</string>
+	<key>IDESourceControlProjectRelativeInstallPathDictionary</key>
+	<dict>
+		<key>31D19277-93C1-4B36-8D47-39BAE008E6C8</key>
+		<string>../../../../../Documents/Arduino/Libraries/AdaFruit_NeoPixel</string>
+		<key>581ADA72-D596-47B7-96A2-A3DFDF0C6864</key>
+		<string>../../../../../Documents/Arduino/Libraries/Canbus</string>
+		<key>7B373F31-FC1C-4563-BCFF-DFA44AB4C0D0</key>
+		<string>../../../../../Documents/Arduino/Libraries/SD</string>
+		<key>7B5D90EB-E67D-4F4F-B891-A4D47362BD97</key>
+		<string>../..</string>
+		<key>B998CEF2-BA3A-4B9A-861D-23AB57E81EA4</key>
+		<string>../../Libraries/FastLED</string>
+		<key>BB79AA8B-9CED-45F6-9A86-44FF4DDD236C</key>
+		<string>../../../../../Documents/Arduino/Libraries/RTClib</string>
+		<key>D49A604E-E9FF-408A-B3BC-33F3DEF41E6B</key>
+		<string>../../../../../Documents/Arduino/Libraries/UTFT</string>
+	</dict>
+	<key>IDESourceControlProjectURL</key>
+	<string>https://github.com/corbinstreehouse/LEDDigitalCyrWheel.git</string>
+	<key>IDESourceControlProjectVersion</key>
+	<integer>110</integer>
+	<key>IDESourceControlProjectWCCIdentifier</key>
+	<string>7B5D90EB-E67D-4F4F-B891-A4D47362BD97</string>
+	<key>IDESourceControlProjectWCConfigurations</key>
+	<array>
+		<dict>
+			<key>IDESourceControlRepositoryExtensionIdentifierKey</key>
+			<string>public.vcs.git</string>
+			<key>IDESourceControlWCCIdentifierKey</key>
+			<string>31D19277-93C1-4B36-8D47-39BAE008E6C8</string>
+			<key>IDESourceControlWCCName</key>
+			<string>AdaFruit_NeoPixel</string>
+		</dict>
+		<dict>
+			<key>IDESourceControlRepositoryExtensionIdentifierKey</key>
+			<string>public.vcs.git</string>
+			<key>IDESourceControlWCCIdentifierKey</key>
+			<string>581ADA72-D596-47B7-96A2-A3DFDF0C6864</string>
+			<key>IDESourceControlWCCName</key>
+			<string>Canbus</string>
+		</dict>
+		<dict>
+			<key>IDESourceControlRepositoryExtensionIdentifierKey</key>
+			<string>public.vcs.git</string>
+			<key>IDESourceControlWCCIdentifierKey</key>
+			<string>B998CEF2-BA3A-4B9A-861D-23AB57E81EA4</string>
+			<key>IDESourceControlWCCName</key>
+			<string>FastLED</string>
+		</dict>
+		<dict>
+			<key>IDESourceControlRepositoryExtensionIdentifierKey</key>
+			<string>public.vcs.git</string>
+			<key>IDESourceControlWCCIdentifierKey</key>
+			<string>7B5D90EB-E67D-4F4F-B891-A4D47362BD97</string>
+			<key>IDESourceControlWCCName</key>
+			<string>LEDDigitalCyrWheel</string>
+		</dict>
+		<dict>
+			<key>IDESourceControlRepositoryExtensionIdentifierKey</key>
+			<string>public.vcs.git</string>
+			<key>IDESourceControlWCCIdentifierKey</key>
+			<string>BB79AA8B-9CED-45F6-9A86-44FF4DDD236C</string>
+			<key>IDESourceControlWCCName</key>
+			<string>RTClib</string>
+		</dict>
+		<dict>
+			<key>IDESourceControlRepositoryExtensionIdentifierKey</key>
+			<string>public.vcs.git</string>
+			<key>IDESourceControlWCCIdentifierKey</key>
+			<string>7B373F31-FC1C-4563-BCFF-DFA44AB4C0D0</string>
+			<key>IDESourceControlWCCName</key>
+			<string>SD</string>
+		</dict>
+		<dict>
+			<key>IDESourceControlRepositoryExtensionIdentifierKey</key>
+			<string>public.vcs.git</string>
+			<key>IDESourceControlWCCIdentifierKey</key>
+			<string>D49A604E-E9FF-408A-B3BC-33F3DEF41E6B</string>
+			<key>IDESourceControlWCCName</key>
+			<string>UTFT</string>
+		</dict>
+	</array>
+</dict>
+</plist>
