+++ examples/Web_HelloWorld/HelloWorld.xcodeproj/xcuserdata/corbin.xcuserdatad/xcschemes/xcschememanagement.plist
@@ -1,0 +1,98 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
+<plist version="1.0">
+<dict>
+	<key>SchemeUserState</key>
+	<dict>
+		<key>All.xcscheme</key>
+		<dict>
+			<key>isShown</key>
+			<false/>
+			<key>orderHint</key>
+			<integer>3</integer>
+		</dict>
+		<key>Boards.xcscheme</key>
+		<dict>
+			<key>isShown</key>
+			<false/>
+			<key>orderHint</key>
+			<integer>0</integer>
+		</dict>
+		<key>Build.xcscheme</key>
+		<dict>
+			<key>orderHint</key>
+			<integer>1</integer>
+		</dict>
+		<key>Clean.xcscheme</key>
+		<dict>
+			<key>orderHint</key>
+			<integer>6</integer>
+		</dict>
+		<key>Index.xcscheme</key>
+		<dict>
+			<key>isShown</key>
+			<true/>
+			<key>orderHint</key>
+			<integer>5</integer>
+		</dict>
+		<key>Make.xcscheme</key>
+		<dict>
+			<key>orderHint</key>
+			<integer>7</integer>
+		</dict>
+		<key>Serial.xcscheme</key>
+		<dict>
+			<key>orderHint</key>
+			<integer>2</integer>
+		</dict>
+		<key>Upload.xcscheme</key>
+		<dict>
+			<key>orderHint</key>
+			<integer>4</integer>
+		</dict>
+	</dict>
+	<key>SuppressBuildableAutocreation</key>
+	<dict>
+		<key>9D77CB9C1589994A0089EB0A</key>
+		<dict>
+			<key>primary</key>
+			<true/>
+		</dict>
+		<key>9D77CBA01589994A0089EB0A</key>
+		<dict>
+			<key>primary</key>
+			<true/>
+		</dict>
+		<key>9D77CBA11589994A0089EB0A</key>
+		<dict>
+			<key>primary</key>
+			<true/>
+		</dict>
+		<key>9D77CBA21589994A0089EB0A</key>
+		<dict>
+			<key>primary</key>
+			<true/>
+		</dict>
+		<key>9D77CBA31589994A0089EB0A</key>
+		<dict>
+			<key>primary</key>
+			<true/>
+		</dict>
+		<key>9D77CBA41589994A0089EB0A</key>
+		<dict>
+			<key>primary</key>
+			<true/>
+		</dict>
+		<key>9D77CBA51589994A0089EB0A</key>
+		<dict>
+			<key>primary</key>
+			<true/>
+		</dict>
+		<key>9D77CBA61589994A0089EB0A</key>
+		<dict>
+			<key>primary</key>
+			<true/>
+		</dict>
+	</dict>
+</dict>
+</plist>
